// packages/destination-actions/src/destinations/klime/group/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Payload {
  /**
   * Unique identifier for the event
   */
  messageId: string
  /**
   * Unique identifier for the group/organization
   */
  groupId: string
  /**
   * User to associate with the group
   */
  userId?: string
  /**
   * Group attributes (name, plan, industry, etc.)
   */
  traits?: {
    [k: string]: unknown
  }
  /**
   * When the event occurred (ISO 8601)
   */
  timestamp?: string | number
  /**
   * Contextual information about the event
   */
  context?: {
    [k: string]: unknown
  }
  /**
   * When enabled, events are sent in batches.
   */
  enable_batching: boolean
  /**
   * Maximum number of events per batch request.
   */
  batch_size?: number
}